+++ Sources/BaseFoundationiOS/Protocols/DataSources/FechaDataSource.swift
@@ -1,0 +1,12 @@
+//
+//  FechaDataSource.swift
+//
+//
+//  Created by pablo borquez on 29/06/24.
+//
+
+import Foundation
+
+public protocol FechaDataSource {
+    var fecha: Date { get set }
+}